# openapi3.yaml
openapi: 3.0.1
info:
  title: osm sync tracker
  description: Api to update on the progress of a cycle of entities sync
  license:
    name: MIT
  version: 3.5.0
...
paths:
  /sync:
    get:
      operationId: getSyncs
      tags:
        - sync
      summary: Get syncs by filter
      parameters:
        - in: query
          name: status
          schema:
            type: array
            items:
              $ref: '#/components/schemas/status'
        - in: query
          name: geometry_type
          schema:
            type: array
            items:
              $ref: '#/components/schemas/geometryType'
        - in: query
          name: layer_id
          schema:
            type: array
            items:
              $ref: '#/components/schemas/layerId'
        - in: query
          name: is_full
          schema:
            type: boolean
        - in: query
          name: is_rerun
          schema:
            type: boolean
      responses:
        '200':
          description: all the syncs matching provided filter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/extendedSync'
        '400':
          $ref: '#/components/responses/BadRequest'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
    post:
      summary: Register a new sync
      requestBody:
        description: start tracking a new synchronization
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/sync'
                - type: object
                  required:
                    - id
                    - startDate
                    - layerId
                    - isFull
                    - dumpDate
                    - geometryType
        required: true
      operationId: postSync
      tags:
        - sync
      responses:
        '201':
          description: created
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /sync/latest:
    get:
      summary: Get the latest sync based on start date
      parameters:
        - in: query
          name: layerId
          schema:
            $ref: '#/components/schemas/layerId'
          required: true
        - in: query
          name: geometryType
          schema:
            $ref: '#/components/schemas/geometryType'
          required: true
      operationId: getLatestSync
      tags:
        - sync
      responses:
        '200':
          description: the latest sync for the requsted layer and geometry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sync'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /sync/closure:
    post:
      summary: creates closure jobs for selected syncs
      description: closure jobs will attempt to close the sync
      requestBody:
        description: the sync ids to create closure jobs to
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/uniqueId'
      operationId: postSyncsClosure
      tags:
        - closure
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/BadRequest'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /sync/{syncId}:
    patch:
      summary: Update the sync
      parameters:
        - $ref: '#/components/parameters/syncId'
      requestBody:
        description: the propeties to update in the sync
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateSync'
        required: true
      operationId: patchSync
      tags:
        - sync
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
...
  /sync/{syncId}/rerun:
    post:
      summary: rerun a sync if needed
      parameters:
        - $ref: '#/components/parameters/syncId'
      requestBody:
        description: the rerun id to be created and its start date
        content:
          application/json:
            schema:
              type: object
              properties:
                rerunId:
                  $ref: '#/components/schemas/uniqueId'
                startDate:
                  type: string
                  format: date-time
                shouldRerunNotSynced:
                  type: boolean
              required:
                - rerunId
                - startDate
      operationId: rerunSync
      tags:
        - sync
      responses:
        '200':
          description: the rerun was not created due to the sync closing successfuly
        '201':
          description: created
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnexpectedError:
      description: Unexpected Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    syncId:
      in: path
      name: syncId
      schema:
        $ref: '#/components/schemas/uniqueId'
      required: true
...
  schemas:
    status:
      type: string
      enum:
        - inprogress
        - completed
        - failed
    geometryType:
      type: string
      enum:
        - point
        - linestring
        - polygon
    Error:
      type: object
      properties:
        message:
          type: string
    uniqueId:
      type: string
      format: uuid
...
    layerId:
      type: integer
    sync:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/uniqueId'
        isFull:
          type: boolean
        startDate:
          type: string
          format: date-time
        dumpDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/status'
        layerId:
          $ref: '#/components/schemas/layerId'
        totalFiles:
          type: integer
        geometryType:
          $ref: '#/components/schemas/geometryType'
        metadata:
          $ref: '#/components/schemas/syncMetadata'
    extendedSync:
      allOf:
        - $ref: '#/components/schemas/sync'
        - type: object
          properties:
            baseSyncId:
              $ref: '#/components/schemas/uniqueId'
            runNumber:
              type: integer
    updateSync:
      type: object
      additionalProperties: false
      properties:
        startDate:
          type: string
          format: date-time
        dumpDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/status'
        layerId:
          $ref: '#/components/schemas/layerId'
        totalFiles:
          type: integer
        geometryType:
          $ref: '#/components/schemas/geometryType'
        metadata:
          $ref: '#/components/schemas/syncMetadata'
...
    syncMetadata:
      type: object
      additionalProperties: true